ml-models/requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.0.74
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
boto3==1.28.25
python-dotenv==1.0.0
h5py==3.9.0
joblib==1.3.2

